LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/perspective-expected.txt


PASS perspective responsive to style changes
PASS perspective clamped to 0px on keyframes
FAIL perspective responsive to inherited changes assert_equals: expected "80px" but got "none"
